requirements.txt
opencv-python>=4.5.0
qrcode>=7.3.1
Pillow>=8.0.0
numpy>=1.19.0
pytz>=2021.1
python-dateutil>=2.8.2 